stego/requirements.txt
matplotlib>=3.8.4
numpy>=1.26.4
opencv_python>=4.9.0.80
pandas>=2.2.2
Pillow>=10.3.0
